backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart
pydantic>=2.0

# PDF parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# NLP / Embeddings
sentence-transformers>=2.6.0
scikit-learn>=1.4.0
numpy>=1.26.0
faiss-cpu>=1.8.0

# LLM (Anthropic Claude — for real streaming answers)
google-generativeai >=0.1.0